+++ packages/builder/builder-shared/src/fs.ts
@@ -1,4 +1,9 @@
-import { CSS_MODULE_REGEX, CSS_REGEX, NODE_MODULES_REGEX } from './constants';
+import {
+  CSS_MODULE_REGEX,
+  CSS_REGEX,
+  MODULE_PATH_REGEX,
+  NODE_MODULES_REGEX,
+} from './constants';
 
 export async function isFileExists(file: string) {
   const { promises, constants } = await import('fs');
@@ -12,3 +17,18 @@ export const isNodeModulesCss = (path: string) =>
   NODE_MODULES_REGEX.test(path) &&
   CSS_REGEX.test(path) &&
   !CSS_MODULE_REGEX.test(path);
+
+export function getPackageNameFromModulePath(modulePath: string) {
+  const handleModuleContext = modulePath?.match(MODULE_PATH_REGEX);
+
+  if (!handleModuleContext) {
+    return false;
+  }
+
+  const [, , scope, name] = handleModuleContext;
+  const packageName = ['npm', (scope ?? '').replace('@', ''), name]
+    .filter(Boolean)
+    .join('.');
+
+  return packageName;
+}
