+++ packages/builder/builder-webpack-provider/src/stub/utils.ts
@@ -1,5 +1,6 @@
-import type { GlobbyOptions } from '@modern-js/utils';
 import fs, { PathLike } from 'fs';
+import type { GlobbyOptions } from '@modern-js/utils';
+import type webpack from 'webpack';
 
 export interface GlobContentJSONOptions extends GlobbyOptions {
   maxSize?: number;
@@ -27,3 +28,46 @@ export const globContentJSON = async (
 
 export const filenameToGlobExpr = (file: PathLike) =>
   fs.statSync(file).isDirectory() ? `${file}/**/*` : file;
+
+/**
+ * Check if a file handled by specific loader.
+ * @author yangxingyuan
+ * @param {Configuration} config - The webpack config.
+ * @param {string} loader - The name of loader.
+ * @param {string}  testFile - The name of test file that will be handled by webpack.
+ * @returns {boolean} The result of the match.
+ */
+export function matchLoader({
+  config,
+  loader,
+  testFile,
+}: {
+  config: webpack.Configuration;
+  loader: string;
+  testFile: string;
+}): boolean {
+  if (!config.module?.rules) {
+    return false;
+  }
+  return config.module.rules.some(rule => {
+    if (
+      typeof rule === 'object' &&
+      rule.test &&
+      rule.test instanceof RegExp &&
+      rule.test.test(testFile)
+    ) {
+      return (
+        Array.isArray(rule.use) &&
+        rule.use.some(useOptions => {
+          if (typeof useOptions === 'object' && useOptions !== null) {
+            return useOptions.loader?.includes(loader);
+          } else if (typeof useOptions === 'string') {
+            return useOptions.includes(loader);
+          }
+          return false;
+        })
+      );
+    }
+    return false;
+  });
+}
